+++ module/applications/_module.mjs
@@ -7,3 +7,4 @@ export * as components from "./components/_index.mjs";
 export * as dialogs from "./dialogs/_module.mjs";
 export * from "./momentum.mjs";
 export * as sheets from "./sheets/_module.mjs";
+export * as ux from "./ux/_module.mjs";